requirements.txt
# Discord Bot
discord.py>=2.3.0
python-dotenv>=1.0.0

# API Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Proxmox API
proxmoxer>=2.0.1
requests>=2.31.0

# WireGuard
wgconfig>=0.2.2

# Patreon API
patreon>=0.5.0

# Utilities
aiohttp>=3.9.0
python-dateutil>=2.8.2
cryptography>=42.0.0
